+++ clients/web3-wallet/examples/rest-api/TradingApi/buildSwapTransaction.ts
@@ -14,7 +14,6 @@ async function buildSwapTransaction() {
             amount: '1000000',
             fromTokenAddress: '0x55d398326f99059fF775485246999027B3197955',
             toTokenAddress: '0x8AC76a51cc950d9822D68b83fE1Ad97B32Cd580d',
-            slippagePercent: '0.5',
             userWalletAddress: '0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045',
             quoteId: 'a1b2c3d4e5f64a8b9c0d1e2f3a4b5c6d',
         });